+++ gui/src/message/mod.rs
@@ -1,0 +1,5 @@
+pub mod message_data;
+pub mod message_row;
+
+pub use message_data::MessageObject;
+pub use message_row::MessageRow;
