+++ gui/src/user/mod.rs
@@ -1,0 +1,5 @@
+pub mod user_data;
+pub mod user_row;
+
+pub use user_data::UserObject;
+pub use user_row::UserRow;
